services/delivery-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
paho-mqtt==2.1.0
pydantic-settings==2.6.1